update PermitRowOptions component to handle both IDIR and BCeID users, remove IDIRPermitSearchRowActions component, implement permissions matrix check for relevant components
