+++ frontend/src/features/idir/search/components/IDIRPermitSearchResults.tsx
@@ -18,7 +18,6 @@ import { PermitListItem } from "../../../permits/types/permit";
 import { getPermitDataBySearch } from "../api/idirSearch";
 import { PermitSearchResultColumnDef } from "../table/PermitSearchResultColumnDef";
 import { PERMIT_ACTION_ORIGINS, SearchFields } from "../types/types";
-import { IDIRPermitSearchRowActions } from "./IDIRPermitSearchRowActions";
 import {
   defaultTableInitialStateOptions,
   defaultTableOptions,
@@ -29,6 +28,7 @@ import { ERROR_ROUTES } from "../../../../routes/constants";
 import { VEHICLES_URL } from "../../../../common/apiManager/endpoints/endpoints";
 import { httpGETRequest } from "../../../../common/apiManager/httpRequestHandler";
 import { useSetCompanyHandler } from "../helpers/useSetCompanyHandler";
+import { PermitRowOptions } from "../../../permits/components/permit-list/PermitRowOptions";
 
 /**
  * Function to decide whether to show row actions icon or not.
@@ -205,11 +205,10 @@ export const IDIRPermitSearchResults = memo(
           if (shouldShowRowActions(idirUserDetails?.userRole)) {
             return (
               <Box className="idir-search-results__row-actions">
-                <IDIRPermitSearchRowActions
+                <PermitRowOptions
                   isPermitInactive={isInactive}
                   permitNumber={row.original.permitNumber}
                   permitId={row.original.permitId}
-                  userRole={idirUserDetails?.userRole}
                   companyId={row.original.companyId}
                   permitActionOrigin={PERMIT_ACTION_ORIGINS.GLOBAL_SEARCH}
                 />
