+++ frontend/src/features/permits/components/permit-list/PermitRowOptions.tsx
@@ -1,60 +1,232 @@
-import { OnRouteBCTableRowActions } from "../../../../common/components/table/OnRouteBCTableRowActions";
-import { viewReceiptPdf } from "../../helpers/permitPDFHelper";
+import { useContext, useState } from "react";
+import { useNavigate } from "react-router-dom";
 
-const PERMIT_ACTION_OPTION_TYPES = {
+import { OnRouteBCTableRowActions } from "../../../../common/components/table/OnRouteBCTableRowActions";
+import PermitResendDialog from "../../../idir/search/components/PermitResendDialog";
+import { viewReceiptPdf } from "../../helpers/permitPDFHelper";
+import * as routes from "../../../../routes/constants";
+import { useResendPermit } from "../../hooks/hooks";
+import { SnackBarContext } from "../../../../App";
+import { EmailNotificationType } from "../../types/EmailNotificationType";
+import { useAttemptAmend } from "../../hooks/useAttemptAmend";
+import { UnfinishedAmendModal } from "../../pages/Amend/components/modal/UnfinishedAmendModal";
+import { getDefaultRequiredVal } from "../../../../common/helpers/util";
+import { PermitActionOrigin } from "../../../idir/search/types/types";
+import { usePermissionMatrix } from "../../../../common/authentication/PermissionMatrix";
+
+const PERMIT_ACTION_TYPES = {
+  RESEND: "resend",
   VIEW_RECEIPT: "viewReceipt",
+  AMEND: "amend",
+  VOID_REVOKE: "voidRevoke",
 } as const;
 
-type PermitActionOptionType =
-  (typeof PERMIT_ACTION_OPTION_TYPES)[keyof typeof PERMIT_ACTION_OPTION_TYPES];
+type PermitActionType =
+  (typeof PERMIT_ACTION_TYPES)[keyof typeof PERMIT_ACTION_TYPES];
 
-const getOptionLabel = (optionType: PermitActionOptionType): string => {
-  if (optionType === PERMIT_ACTION_OPTION_TYPES.VIEW_RECEIPT) {
-    return "View Receipt";
+const permitActionLabel = (actionType: PermitActionType) => {
+  switch (actionType) {
+    case PERMIT_ACTION_TYPES.RESEND:
+      return "Resend";
+    case PERMIT_ACTION_TYPES.VIEW_RECEIPT:
+      return "View Receipt";
+    case PERMIT_ACTION_TYPES.AMEND:
+      return "Amend";
+    case PERMIT_ACTION_TYPES.VOID_REVOKE:
+      return "Void/Revoke";
+    default:
+      return "";
   }
-
-  return "";
 };
 
-const ALL_OPTIONS = [
-  {
-    label: getOptionLabel(PERMIT_ACTION_OPTION_TYPES.VIEW_RECEIPT),
-    value: PERMIT_ACTION_OPTION_TYPES.VIEW_RECEIPT,
-  },
-];
-
-const getOptions = (isExpired: boolean) => {
-  if (isExpired) {
-    return ALL_OPTIONS;
-  }
-  return ALL_OPTIONS;
-};
-
+/**
+ * Component for row actions on IDIR Search Permit.
+ */
 export const PermitRowOptions = ({
-  isExpired,
-  companyId,
   permitId,
-  onDocumentUnavailable,
+  isPermitInactive,
+  permitNumber,
+  companyId,
+  permitActionOrigin,
 }: {
-  isExpired: boolean;
-  companyId: number;
-  permitId: string;
-  onDocumentUnavailable: () => void;
-}) => {
   /**
-   * Action handler upon a select event.
-   * @param selectedOption The option that was selected.
+   * The permit id.
    */
-  const onSelectOptionCallback = (selectedOption: string) => {
-    if (selectedOption === PERMIT_ACTION_OPTION_TYPES.VIEW_RECEIPT) {
-      viewReceiptPdf(companyId, permitId, () => onDocumentUnavailable());
+  permitId: string;
+  /**
+   * Is the permit inactive (voided/superseded/revoked) or expired?
+   */
+  isPermitInactive: boolean;
+  /**
+   * The permit number
+   */
+  permitNumber: string;
+  /**
+   * The company id
+   */
+  companyId: number;
+  /**
+   * The application location from where the permit action (amend / void / revoke) originated
+   */
+  permitActionOrigin: PermitActionOrigin;
+}) => {
+  const [openResendDialog, setOpenResendDialog] = useState<boolean>(false);
+  const navigate = useNavigate();
+  const resendPermitMutation = useResendPermit();
+  const { setSnackBar } = useContext(SnackBarContext);
+
+  const {
+    choosePermitToAmend,
+    showUnfinishedModal,
+    existingAmendmentApplication,
+    handleCloseModal,
+    handleStartNewAmendment,
+    handleContinueAmendment,
+  } = useAttemptAmend(permitActionOrigin);
+
+  const existingAmendmentCreatedBy = getDefaultRequiredVal(
+    "",
+    existingAmendmentApplication?.applicant,
+  );
+
+  /**
+   * Function to handle user selection from the options.
+   * @param selectedOption The selected option as a string.
+   */
+  const onSelectOption = (selectedOption: string) => {
+    if (selectedOption === PERMIT_ACTION_TYPES.RESEND) {
+      setOpenResendDialog(() => true);
+    } else if (selectedOption === PERMIT_ACTION_TYPES.VIEW_RECEIPT) {
+      viewReceiptPdf(companyId, permitId, () =>
+        navigate(routes.ERROR_ROUTES.DOCUMENT_UNAVAILABLE),
+      );
+    } else if (selectedOption === PERMIT_ACTION_TYPES.VOID_REVOKE) {
+      navigate(`${routes.PERMITS_ROUTES.VOID(companyId, permitId)}`);
+    } else if (selectedOption === PERMIT_ACTION_TYPES.AMEND) {
+      // Sets the companyId and permitId of the permit to be amended,
+      // which will in turn look for any existing associated amendment applications,
+      // which is used to show info in the modal (or not show the modal at all)
+      choosePermitToAmend(companyId, permitId);
     }
   };
 
+  const handleResend = async (
+    permitId: string,
+    email: string,
+    notificationTypes: EmailNotificationType[],
+  ) => {
+    const response = await resendPermitMutation.mutateAsync({
+      permitId,
+      email,
+      notificationTypes,
+    });
+
+    setOpenResendDialog(false);
+    if (response.status === 201) {
+      setSnackBar({
+        showSnackbar: true,
+        setShowSnackbar: () => true,
+        message: "Successfully sent",
+        alertType: "success",
+      });
+    } else {
+      navigate(routes.ERROR_ROUTES.UNEXPECTED, {
+        state: { correlationId: response.headers["x-correlation-id"] },
+      });
+    }
+  };
+
+  const canResendPermit = usePermissionMatrix({
+    permissionMatrixKeys: {
+      permissionMatrixFeatureKey: "GLOBAL_SEARCH",
+      permissionMatrixFunctionKey: "RESEND_PERMIT",
+    },
+  });
+
+  const canViewPermitReceipt = usePermissionMatrix({
+    permissionMatrixKeys: {
+      permissionMatrixFeatureKey: "MANAGE_PERMITS",
+      permissionMatrixFunctionKey: "VIEW_PERMIT_RECEIPT",
+    },
+  });
+
+  const canAmendPermit = usePermissionMatrix({
+    permissionMatrixKeys: {
+      permissionMatrixFeatureKey: "GLOBAL_SEARCH",
+      permissionMatrixFunctionKey: "AMEND_PERMIT",
+    },
+  });
+
+  const canVoidPermit = usePermissionMatrix({
+    permissionMatrixKeys: {
+      permissionMatrixFeatureKey: "GLOBAL_SEARCH",
+      permissionMatrixFunctionKey: "VOID_PERMIT",
+    },
+  });
+
+  interface PermitAction {
+    action: PermitActionType;
+    isAuthorized: (isExpired: boolean) => boolean;
+  }
+
+  const PERMIT_ACTIONS: PermitAction[] = [
+    {
+      action: PERMIT_ACTION_TYPES.RESEND,
+      isAuthorized: () => canResendPermit,
+    },
+    {
+      action: PERMIT_ACTION_TYPES.VIEW_RECEIPT,
+      isAuthorized: () => canViewPermitReceipt,
+    },
+    {
+      action: PERMIT_ACTION_TYPES.AMEND,
+      isAuthorized: (isExpired: boolean) => !isExpired && canAmendPermit,
+    },
+    {
+      action: PERMIT_ACTION_TYPES.VOID_REVOKE,
+      isAuthorized: (isExpired: boolean) => !isExpired && canVoidPermit,
+    },
+  ];
+
+  /**
+   * Returns options for the row actions.
+   * @param isExpired Has the permit expired?
+   * @returns Action options that can be performed for the permit.
+   */
+  const getOptions = (isExpired: boolean) => {
+    return PERMIT_ACTIONS.filter((action) =>
+      action.isAuthorized(isExpired),
+    ).map(({ action }) => ({
+      label: permitActionLabel(action),
+      value: action,
+    }));
+  };
+
   return (
-    <OnRouteBCTableRowActions
-      onSelectOption={onSelectOptionCallback}
-      options={getOptions(isExpired)}
-    />
+    <>
+      <OnRouteBCTableRowActions
+        onSelectOption={onSelectOption}
+        options={getOptions(isPermitInactive)}
+        key={`idir-search-row-${permitNumber}`}
+      />
+
+      <PermitResendDialog
+        shouldOpen={openResendDialog}
+        onCancel={() => setOpenResendDialog(false)}
+        onResend={handleResend}
+        companyId={companyId}
+        permitId={permitId}
+        permitNumber={permitNumber}
+      />
+
+      <UnfinishedAmendModal
+        shouldOpen={showUnfinishedModal}
+        issuedPermitNumber={permitNumber}
+        unfinishedAmendmentCreatedBy={existingAmendmentCreatedBy}
+        onCancel={handleCloseModal}
+        onStartNewAmendment={handleStartNewAmendment}
+        onContinueAmendment={handleContinueAmendment}
+      />
+    </>
   );
 };
