+++ frontend/src/features/permits/pages/Void/VoidPermit.tsx
@@ -1,5 +1,5 @@
 import { Navigate, useNavigate, useParams } from "react-router-dom";
-import { useState, useEffect, useContext, useMemo } from "react";
+import { useState, useEffect, useMemo } from "react";
 
 import { VoidPermitForm } from "./components/VoidPermitForm";
 import { usePermitDetailsQuery } from "../../hooks/hooks";
@@ -10,17 +10,19 @@ import { VoidPermitContext } from "./context/VoidPermitContext";
 import { ERROR_ROUTES, IDIR_ROUTES } from "../../../../routes/constants";
 import { VoidPermitFormData } from "./types/VoidPermit";
 import { FinishVoid } from "./FinishVoid";
-import OnRouteBCContext from "../../../../common/authentication/OnRouteBCContext";
-import { USER_ROLE } from "../../../../common/authentication/types";
 import { isPermitInactive } from "../../types/PermitStatus";
 import { Permit } from "../../types/permit";
-import { applyWhenNotNullable, getDefaultRequiredVal } from "../../../../common/helpers/util";
+import {
+  applyWhenNotNullable,
+  getDefaultRequiredVal,
+} from "../../../../common/helpers/util";
 import { Breadcrumb } from "../../../../common/components/breadcrumb/Breadcrumb";
 import { hasPermitExpired } from "../../helpers/permitState";
 import {
   SEARCH_BY_FILTERS,
   SEARCH_ENTITIES,
 } from "../../../idir/search/types/types";
+import { usePermissionMatrix } from "../../../../common/authentication/PermissionMatrix";
 
 const searchRoute =
   `${IDIR_ROUTES.SEARCH_RESULTS}?searchEntity=${SEARCH_ENTITIES.PERMIT}` +
@@ -35,19 +37,17 @@ const isVoidable = (permit: Permit) => {
 
 export const VoidPermit = () => {
   const navigate = useNavigate();
-  const {
-    permitId: permitIdParam,
-    companyId: companyIdParam,
-  } = useParams();
+  const { permitId: permitIdParam, companyId: companyIdParam } = useParams();
 
-  const companyId: number = applyWhenNotNullable(id => Number(id), companyIdParam, 0);
+  const companyId: number = applyWhenNotNullable(
+    (id) => Number(id),
+    companyIdParam,
+    0,
+  );
   const permitId = getDefaultRequiredVal("", permitIdParam);
   const [currentLink, setCurrentLink] = useState(0);
   const getBannerText = () =>
     currentLink === 0 ? "Void Permit" : "Finish Voiding";
-
-  // Must be SYSADMIN to access this page
-  const { idirUserDetails } = useContext(OnRouteBCContext);
 
   const permitQuery = usePermitDetailsQuery(companyId, permitId);
   const permit = permitQuery.data;
@@ -116,8 +116,15 @@ export const VoidPermit = () => {
     [voidPermitData],
   );
 
+  const canVoidPermit = usePermissionMatrix({
+    permissionMatrixKeys: {
+      permissionMatrixFeatureKey: "GLOBAL_SEARCH",
+      permissionMatrixFunctionKey: "VOID_PERMIT",
+    },
+  });
+
   // If user is not SYSADMIN, show unauthorized page
-  if (idirUserDetails?.userRole !== USER_ROLE.SYSTEM_ADMINISTRATOR) {
+  if (!canVoidPermit) {
     return <Navigate to={ERROR_ROUTES.UNAUTHORIZED} />;
   }
